+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Non-weighted flow statistics are hidden in a collapsed section of the report
+
 ### Deprecated
 
 ### Removed